refactor: move `get_allow_fullscreen` to `fullscreen` subfolder
