+++ components/script/dom/document/document.rs
@@ -4698,26 +4698,6 @@ impl Document {
         self.fullscreen_element.set(element);
     }
 
-    pub(crate) fn get_allow_fullscreen(&self) -> bool {
-        // https://html.spec.whatwg.org/multipage/#allowed-to-use
-        match self.browsing_context() {
-            // Step 1
-            None => false,
-            Some(_) => {
-                // Step 2
-                let window = self.window();
-                if window.is_top_level() {
-                    true
-                } else {
-                    // Step 3
-                    window
-                        .GetFrameElement()
-                        .is_some_and(|el| el.has_attribute(&local_name!("allowfullscreen")))
-                }
-            },
-        }
-    }
-
     fn reset_form_owner_for_listeners(&self, id: &Atom, can_gc: CanGc) {
         let map = self.form_id_listener_map.borrow();
         if let Some(listeners) = map.get(id) {
