+++ components/script/dom/fullscreen/fullscreen.rs
@@ -1,0 +1,26 @@
+use html5ever::local_name;
+
+use crate::dom::bindings::codegen::Bindings::WindowBinding::WindowMethods;
+use crate::dom::document::document::Document;
+
+impl Document {
+    pub(crate) fn get_allow_fullscreen(&self) -> bool {
+        // https://html.spec.whatwg.org/multipage/#allowed-to-use
+        match self.browsing_context() {
+            // Step 1
+            None => false,
+            Some(_) => {
+                // Step 2
+                let window = self.window();
+                if window.is_top_level() {
+                    true
+                } else {
+                    // Step 3
+                    window
+                        .GetFrameElement()
+                        .is_some_and(|el| el.has_attribute(&local_name!("allowfullscreen")))
+                }
+            },
+        }
+    }
+}
